Gui upgrade 02
1. GuiRoot. currentScreen() closeScreen()...
2. Gui.addLayoutorWrapChildren()
3. Gui.addLayoutorAlignParentRR/LTRB
4. GuiProfileVisual.java: choose display section.
5. GuiWindow
6. GuiMenubar, GuiMenu OP
7. Gui.indexOfGui(), lastIndexOfGui(), getChildren()
8. Gui.addOnDraggingListener new way
9. new concept in GuiText - just a indie child, should not too much extends

+++ src/outskirts/client/render/Framebuffer.java
@@ -36,7 +36,7 @@ public class Framebuffer {
 
     public static void bindMainFramebuffer() { // or should be isolate with this Object, as a static method bindMainFramebuffer() ..?
         glBindFramebuffer(GL_FRAMEBUFFER, 0);
-        glViewport(0, 0, (int)Outskirts.toFramebufferCoords(Outskirts.getWidth()), (int)Outskirts.toFramebufferCoords(Outskirts.getHeight()));
+        glViewport(0, 0, Outskirts.toFramebufferCoords(Outskirts.getWidth()), Outskirts.toFramebufferCoords(Outskirts.getHeight()));
     }
 
     //
